Extract classes from GherkinImporter
